refactor: updated structs names to better reflect their role
separated header and body strategies to allow for more granular control over which strategy to use in some cases (optional content)

updated tests and README.md file to match renamed structs
