+++ Sources/Postie/Strategies/Body/DefaultBodyStrategy.swift
@@ -1,0 +1,8 @@
+public struct DefaultBodyStrategy: ResponseBodyDecodingStrategy {
+
+    public static var allowsEmptyContent: Bool {
+        return false
+    }
+
+    public static var statusCode: Int?
+}
