+++ Sources/Postie/Strategies/Body/OptionalContentStrategy.swift
@@ -1,0 +1,8 @@
+public class OptionalContentStrategy: ResponseBodyDecodingStrategy {
+
+    public static var allowsEmptyContent: Bool {
+        return statusCode == 204
+    }
+
+    public static var statusCode: Int?
+}
