+++ Sources/Postie/Strategies/Body/ResponseBodyDecodingStrategy.swift
@@ -1,5 +1,8 @@
 public protocol ResponseBodyDecodingStrategy {
 
+    /// Indicates wether the decoding should fail on empty body or not.
+    static var allowsEmptyContent: Bool { get }
+
     ///
     /// Use this property in a custom strategy implementation (example below) to check against the response's `statusCode` and determine wether or not it should fail when receiving empty data.
     ///
@@ -8,7 +11,7 @@ public protocol ResponseBodyDecodingStrategy {
     ///
     ///     public static var statusCode: Int?
     ///
-    ///     public static var allowsEmptyBody: Bool {
+    ///     public static var allowsEmptyContent: Bool {
     ///         guard let statusCode = statusCode else {
     ///             return false
     ///         }
@@ -17,7 +20,4 @@ public protocol ResponseBodyDecodingStrategy {
     /// }
     ///
     static var statusCode: Int? { get set }
-
-    /// Indicates wether the decoding should fail on empty body or not.
-    static var allowsEmptyBody: Bool { get }
 }
