+++ Sources/Postie/Strategies/Headers/DefaultHeaderStrategy.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+public struct DefaultHeaderStrategy: ResponseHeaderDecodingStrategy {
+
+    public typealias RawValue = String
+
+    public static func decode(decoder: Decoder) throws -> RawValue {
+        guard let key = decoder.codingPath.last?.stringValue else {
+            throw ResponseHeaderDecodingError.missingCodingKey
+        }
+        // Check if the decoder is response decoder, otherwise fall back to default decoding logic
+        guard let responseDecoding = decoder as? ResponseDecoding else {
+            return try RawValue(from: decoder)
+        }
+        // Transform dash separator to camelCase
+        guard let value = responseDecoding.valueForHeaderCaseInsensitive(key) else {
+            throw DecodingError.valueNotFound(RawValue.self, DecodingError.Context(
+                codingPath: decoder.codingPath,
+                debugDescription: "Missing value for case-insensitive header key: \(key)"
+            ))
+        }
+        return value
+    }
+}
