+++ Sources/Postie/Strategies/Headers/ResponseHeaderDecodingStrategy.swift
@@ -1,0 +1,9 @@
+public protocol ResponseHeaderDecodingStrategy {
+    associatedtype RawValue: Codable
+
+    static func decode(decoder: Decoder) throws -> RawValue
+}
+
+enum ResponseHeaderDecodingError: Error {
+    case missingCodingKey
+}
