--- Sources/Postie/Strategies/ResponseHeaderDecodingStrategy.swift
@@ -1,9 +1,0 @@
-public protocol ResponseHeaderDecodingStrategy {
-  associatedtype RawValue: Codable
-
-  static func decode(decoder: Decoder) throws -> RawValue
-}
-
-enum ResponseHeaderDecodingError: Error {
-  case missingCodingKey
-}
